pagamentos: carga horaria semanal em hora e minuto com conversao para decimal
O formulario de hora extra passa a receber a carga horaria semanal em dois
campos (hora e minuto) e exibe abaixo deles uma dica com o valor ja
convertido para hora decimal (ex: 4h30 = 4,5). Esse valor convertido e o
que e gravado, exportado na coluna Horas da planilha e exibido na consulta
e no modal (filtro decimal_hours, padrao pt-BR).

- forms.py: campos weekly_workload_hours/minutes + workload_decimal()
  (2 casas, arredondamento comercial)
- models.py + migracao e7d5a9c3b1f8: weekly_workload Integer -> Numeric(5,2);
  lancamentos antigos (inteiros) continuam validos
- payments.py: create/edit gravam o decimal (total > 0 validado com flash),
  GET do edit reparte o decimal em hora/minuto (4.33 -> 4h20) e a
  exportacao escreve o decimal como numero
- form_overtime.html: correcao do Mes Base que exibia "Janeiro" no lugar do
  mes corrente — o Tom Select do base.html captura o valor antes do script
  da pagina; agora sincroniza via tomSelectSync
- testes: conversao no create/edit, rejeicao de minutos>59 e total=0,
  exportacao com 4.5 e reparticao na edicao

+++ app/models.py
@@ -281,7 +281,9 @@ class TeacherOvertimePay(db.Model):
     teacher_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
     unity_id = db.Column(db.Integer, db.ForeignKey('unities.id'), nullable=True, index=True)
     teaching_level = db.Column(db.String(50), nullable=False) # E.g., 'Técnico', 'Superior'
-    weekly_workload = db.Column(db.Integer, nullable=False)
+    # Hora decimal com 2 casas (ex: 4.5 = 4h30) — o formulário recebe hora e
+    # minuto separados e grava o valor já convertido.
+    weekly_workload = db.Column(db.Numeric(5, 2), nullable=False)
     hourly_value = db.Column(db.Numeric(10, 2), nullable=False) # 10 dígitos no total, 2 decimais
     budget_code = db.Column(db.String(18), nullable=False)
     shift = db.Column(db.String(50), nullable=False) # E.g., 'Matutino', 'Vespertino', 'Noturno'
